backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["LucasFDutra <lucasfelipedutra@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
psycopg2 = "^2.8.5"
python-dotenv = "^0.14.0"
flask-cors = "^3.0.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
dev = "scripts:run_dev"
migration_up = "scripts:up_migration"
migration_down = "scripts:down_migration"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
